TIP-1185: Fix specs for query builders
